fix(runtime): split GitHub control-flow parsing out of oversized modules
Keep parser security coverage in focused test modules so quality ratchets do not grow already oversized files.
