+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
   "maximum_collected_cases": 16230,
-  "maximum_test_source_lines": 350587
+  "maximum_test_source_lines": 350682
 }
