+++ tests/fixtures/guard-command-corpus/decision-diff-report.json
@@ -20,7 +20,7 @@
       "source-044f1c0e4b5a9e1cedfa7065": "038c2531b51bdb9e777ce27eddd9eb47ba9d2d5d29bb173c917f4383a9a74797",
       "source-05328ef847b0e979fa3bd03c": "64858338b8722344a1a30da5fbeda1cd6325e790d83979e4701229deeb68eda5",
       "source-0568fda2efc7a8044a58a4de": "4584361a7c51cb79073f2c5361a544623c082fffdf303ccbe73b05d843d41dc9",
-      "source-06fdfe7dc0a4051a6134d48d": "0fa255d219fcf6410f5a97066918de40d766433d95881f058e157de58dfde78f",
+      "source-06fdfe7dc0a4051a6134d48d": "86869c438a8c92200220c9bc253583875f2dfc022d2dc86faed5dbaa8cdfecfe",
       "source-07519deb4c03fa6e2ad80126": "19c24d05e666afb1db8087a2ab435713bb7e02082b901e737f4145ce2bbb13af",
       "source-0859591d6f5efaeab51c92af": "562ec5970e2c66507e6f9523a49ed200f92cfb9e6d081dba53be80ce3a9f6ad9",
       "source-087f95a8e656b42f86eab662": "412c8fe443f4a5d7802ac4dfddc64bfe6117ddfca2846e8a6b3a03ed37451097",
@@ -79,7 +79,7 @@
       "source-2e8ccdecc6ece5482a01d405": "70504bd64f5353ee3d3fae082932f407aeb9e5aef63add64c1952c84c6870d59",
       "source-2f280faa31a5e83763dfc77b": "347d1ef23212467d5bb041826e4f87af183b17a711820f00038041a24ead16fd",
       "source-2f536ba0f104e76b74d4ea8e": "8b24ab067420802c13991c18dd28636d4d56b094ce72636dd4419459a9908abd",
-      "source-3025034bc040ba0a2993f316": "fb0fb4a6aaefef64f2a0b8d1d9bb251e66d252ae368a269f5381f4be6efc8e8e",
+      "source-3025034bc040ba0a2993f316": "d92601f091df8c63a4d41f1fdbde0690c24e1c6603a7285f19c5df3771da3308",
       "source-327dd79d0a3617ffa3cab480": "47b3b47e2c605cabb4e1ef3c14ce3be6b978846f3a2f1b08aa4b239b9a14e71a",
       "source-340bdca607709665d6c98dbc": "235bfd8ba820a0c7740f783790a075fed1c3eaa7a6c8e227733face1962b35d8",
       "source-343ab305d578c73898a95d1f": "14d80d4beafc4b02e795b0b3537c12900cbc3f16a6ccd25cfeaa52bfc1c7c720",
@@ -143,6 +143,7 @@
       "source-5c6c23847e3e214f55b45498": "48b89da41282f4643af0f23d48000875a71c039269939af16f33cdaeea8345ec",
       "source-5cffdc3692e0544c89e12f36": "909c988822e0a2816e61cb16f7b797f3507d2437730e2bf311d9e48f0e828a1e",
       "source-5e6923ace9392a7f8893a592": "cc772ae2fde1bc396d290fa2f3326b60bd06a060a462d3c0b47da35c4aeb9eae",
+      "source-5f759a7f40f4284fdead40c7": "2c6129262a056057f84e96d75368abb69d75bf14feda2d452841a0747af85e80",
       "source-5f9b624dfe54e8694f6db1a7": "e4f46e49ca92b342d9449245eb67064b749d13a844934dcc68b83a383cfe4bcd",
       "source-5fa535b4eca33b32fc4d41cd": "21c96c2bfbb655cce88f70c17cad3a8681c9cfc4d87118fbb6685b8d88c6032f",
       "source-5feecbb590c534c10863bb02": "e20cec9b3664f234b7774994c47bd05aa01552d0f3d7279b3fcac27fa0e30910",
